Move Regelbedarf functions to new module.
